requirements.txt
psutil
py-cpuinfo
Pillow
PyMuPDF
yt_dlp
pygame
